+++ src/daf/guild.py
@@ -868,7 +868,7 @@ class AutoGUILD:
         self.guild_query_iter = None
         self.last_guild_join = datetime.min
         self.guild_join_count = 0
-        attributes.write_non_exist(self, "_safe_sem", asyncio.Semaphore(2))
+        attributes.write_non_exist(self, "_safe_sem", asyncio.Semaphore(1))
 
     @property
     def guilds(self) -> List[GUILD]: