+++ src/daf_gui/object.py
@@ -73,7 +73,8 @@ ADDITIONAL_PARAMETER_VALUES = {
 }
 
 DEPRECATION_NOTICES = {
-    daf.AUDIO: [("Youtube streaming", "2.10", "Faster loading times.")],
+    daf.AUDIO: [("daf.dtypes.AUDIO", "2.12", "Replaced with daf.dtypes.FILE")],
+    daf.VoiceMESSAGE: [("daf.dtypes.AUDIO as type for data parameter", "2.12", "Replaced with daf.dtypes.FILE")],
     daf.AutoCHANNEL: [("Interval parameter", "2.11", "Event based refresh.")],
     daf.AutoGUILD: [("Interval parameter", "2.11", "Event based refresh.")],
 }